mmrag/requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
peft
pytorch-lightning
fastapi
uvicorn
omegaconf
nltk
pillow
scikit-learn 